Move the error cleaners to the right file
